+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Command to toggle experimental shaders.
 - Faster Energy Regeneration while sitting.
 - Lantern glow for dropped lanterns.
+- Suggests commands when an invalid one is entered in chat and added Client-side commands to /help.
 
 ### Changed
 - Bats move slower and use a simple proportional controller to maintain altitude